added integration tests
